Improve Roo-mergability
